+++ src/main.js
@@ -88,5 +88,5 @@ targetWindow.MWITools = {
     // Debug utilities available manually via console
     // gameMechanicsAudit,
     // debugEnhancementSpeed,
-    version: '0.4.4'
+    version: '0.4.5'
 };